rename rtcpreceiver into rtpreceiver, rtcpsender into rtpsender
These utilities were originally meant to handle RTCP only, now they are
in charge of every QoS mechanism.
